+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [1.0.0-alpha6] - 2019-02-22
+### Add
+ - Requests per branch is now limited by `shardimage\shardimagephp\auth\Client` class `batchLimit` property. Optional parameter, the client won't send the request if the limit is reached.
 ### Change
  - **[BC BREAK]** If case of error, service will throw response exception, not generating it from the response datas.